style: 전체 여백, 폰트 수정 및 디렉토리 구조 변경
