+++ src/components/sections/CategorySection.tsx
