+++ src/layouts/MainLayout.tsx
@@ -1,14 +1,14 @@
 import React from "react";
-import Navbar from "../components/Navbar";
-import Header from "../components/Header";
+import Navbar from "../components/common/Navbar";
+import Header from "../components/common/Header";
 
 const MainLayout = (props: { children: React.ReactNode }) => {
   return (
-    <>
+    <div className="flex min-h-screen flex-col">
       <Header title="지출 내역 추가" />
-      <main>{props.children}</main>
+      <main className="flex flex-grow flex-col">{props.children}</main>
       <Navbar />
-    </>
+    </div>
   );
 };
